split components out in seperate files
