+++ src/components/create-application/create-application-scrim.tsx
@@ -1,75 +1,22 @@
-import {
-  Button,
-  Checkbox,
-  Divider,
-  Icon,
-  Typography,
-} from '@equinor/eds-core-react';
+import { Button, Divider, Icon, Typography } from '@equinor/eds-core-react';
 import { add } from '@equinor/eds-icons';
 import { useRef, useState } from 'react';
 import { Link, useSearchParams } from 'react-router-dom';
-
-import { routes } from '../../routes';
-import { routeWithParams } from '../../utils/string';
-import {
-  ConfigureApplicationGithub,
-  ConfigureGithubWebhook,
-} from '../configure-application-github';
-import { ScrimPopup } from '../scrim-popup';
-
-import {
-  type ApplicationRegistration,
-  type DeployKeyAndSecret,
-  radixApi,
-  useGetDeployKeyAndSecretQuery,
-  useRegisterApplicationMutation,
-} from '../../store/radix-api';
-import { NewApplyConfigPipelineLink } from '../link/apply-config-pipeline-link';
-import './style.css';
 import useLocalStorage from '../../effects/use-local-storage';
-import { externalUrls } from '../../externalUrls';
-import { pollingInterval } from '../../store/defaults';
-import { ExternalLink } from '../link/external-link';
+import { routes } from '../../routes';
+import type {
+  ApplicationRegistration,
+  DeployKeyAndSecret,
+} from '../../store/radix-api';
+import { routeWithParams } from '../../utils/string';
+import { ConfigureGithubDeploykey } from '../configure-application-github/configure-github-deploykey';
+import { ConfigureGithubWebhook } from '../configure-application-github/configure-github-webhook';
+import { NewApplyConfigPipelineLink } from '../link/apply-config-pipeline-link';
+import { ScrimPopup } from '../scrim-popup';
 import { CreateApplicationForm } from './create-application-form';
+import { UseGithubCIForm } from './use-github-ci-form';
 
-export default function PageCreateApplication() {
-  const [refreshApps] = radixApi.endpoints.showApplications.useLazyQuery({});
-  const [createApp] = useRegisterApplicationMutation();
-  const [appName, setAppName] = useState<string>();
-
-  const { data: secrets } = useGetDeployKeyAndSecretQuery(
-    { appName: appName! },
-    { pollingInterval, skip: !appName }
-  );
-
-  const onCreateApplication = async (
-    applicationRegistration: ApplicationRegistration,
-    acknowledgeWarnings: boolean
-  ) => {
-    const response = await createApp({
-      applicationRegistrationRequest: {
-        applicationRegistration,
-        acknowledgeWarnings,
-      },
-    }).unwrap();
-
-    if (response.applicationRegistration) {
-      setAppName(response.applicationRegistration.name);
-    }
-
-    return response.warnings ?? [];
-  };
-
-  return (
-    <PageCreateApplicationLayout
-      secrets={secrets}
-      onCreateApplication={onCreateApplication}
-      onRefreshApps={() => refreshApps({})}
-    />
-  );
-}
-
-type PageCreateApplicationLayoutProps = {
+type Props = {
   secrets?: DeployKeyAndSecret;
   onRefreshApps: () => unknown;
   onCreateApplication: (
@@ -78,11 +25,11 @@ type PageCreateApplicationLayoutProps = {
   ) => Promise<string[]>;
 };
 
-export function PageCreateApplicationLayout({
+export function CreateApplicationScrim({
   secrets,
   onRefreshApps,
   onCreateApplication,
-}: PageCreateApplicationLayoutProps) {
+}: Props) {
   const [visibleScrim, setVisibleScrim] = useState(false);
   const [useGithub, setUseGithub] = useState<boolean>(true);
   const containerRef = useRef<HTMLDivElement>(null);
@@ -159,10 +106,7 @@ export function PageCreateApplicationLayout({
                 To integrate with GitHub you must add a deploy key and a webhook
               </Typography>
               <Divider />
-              <ConfigureApplicationGithub
-                secrets={secrets}
-                app={registration}
-              />
+              <ConfigureGithubDeploykey secrets={secrets} app={registration} />
               <Divider style={{ width: '100%' }} />
               <Button onClick={() => setNewPage('ci')}>Configure CI</Button>
             </>
@@ -170,34 +114,10 @@ export function PageCreateApplicationLayout({
 
           {page === 'ci' && registration && (
             <>
-              <label htmlFor={'deployOnly'} className="check-input">
-                <Checkbox
-                  id={'deployOnly'}
-                  name="deployOnly"
-                  checked={useGithub}
-                  onChange={(e) => setUseGithub(e.target.checked)}
-                />{' '}
-                <span className="grid grid--gap-small">
-                  <Typography
-                    className="label"
-                    group="input"
-                    variant="text"
-                    token={{ color: 'currentColor' }}
-                  >
-                    Use Radix for CI (build)
-                  </Typography>
-                  <Typography token={{ color: 'currentColor' }}>
-                    Unselect this option if your project is hosted on multiple
-                    repositories and/or requires external control of building.
-                    Radix will no longer need a webhook and will instead deploy
-                    your app through the API/CLI. Read the{' '}
-                    <ExternalLink href={externalUrls.deployOnlyGuide}>
-                      Deployment Guide
-                    </ExternalLink>{' '}
-                    for details.
-                  </Typography>
-                </span>
-              </label>
+              <UseGithubCIForm
+                useGithub={useGithub}
+                setUseGithub={setUseGithub}
+              />
               <Divider style={{ width: '100%' }} />
               {useGithub ? (
                 <Button onClick={() => setNewPage('webhook')}>
